Application/back-end/requirements.txt
gunicorn==23.0.0
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
celery[redis]==5.4.0
celery-singleton==0.3.1
django-allauth[socialaccount]==65.2.0
dj-rest-auth[with_social]==7.0.0
Pillow
requests
